reorient cli as a package.
